backend-moni/pyproject.toml
[project]
name = "backend-moni"
version = "0.1.0"
description = ""
authors = [
    {name = "Gustavo Gonzalez",email = "Gegf1234@Gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "markdown (>=3.7,<4.0)",
    "django-filter (>=25.1,<26.0)",
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
